requirements.txt
google-api-python-client
python-telegram-bot
requests